% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_region.R
\name{find_region}
\alias{find_region}
\title{find_region(gpos=c(10614056,10596790),Ranges=geneRanges)}
\usage{
find_region(gpos, Ranges, ranges.outside = T)
}
\description{
find_region(gpos=c(10614056,10596790),Ranges=geneRanges)
}
